fix: register processor for AccessControlPolicy
