+++ pkg/gateway/cache/methods.go
@@ -2,7 +2,6 @@ package cache
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	discoveryv1 "k8s.io/api/discovery/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gwv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	gwv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
@@ -10,70 +9,8 @@ import (
 	gwpav1alpha1 "github.com/flomesh-io/fsm/pkg/apis/policyattachment/v1alpha1"
 	"github.com/flomesh-io/fsm/pkg/constants"
 
-	mcsv1alpha1 "github.com/flomesh-io/fsm/pkg/apis/multicluster/v1alpha1"
 	gwutils "github.com/flomesh-io/fsm/pkg/gateway/utils"
 )
-
-// Insert inserts an object into the cache
-func (c *GatewayCache) Insert(obj interface{}) bool {
-	p := c.getProcessor(obj)
-	if p != nil {
-		return p.Insert(obj, c)
-	}
-
-	return false
-}
-
-// Delete deletes an object from the cache
-func (c *GatewayCache) Delete(obj interface{}) bool {
-	p := c.getProcessor(obj)
-	if p != nil {
-		return p.Delete(obj, c)
-	}
-
-	return false
-}
-
-//func (c *GatewayCache) WaitForCacheSync(ctx context.Context) bool {
-//	return c.cache.WaitForCacheSync(ctx)
-//}
-
-func (c *GatewayCache) getProcessor(obj interface{}) Processor {
-	switch obj.(type) {
-	//case *corev1.Endpoints:
-	//	return c.processors[EndpointsProcessorType]
-	case *corev1.Service:
-		return c.processors[ServicesProcessorType]
-	case *mcsv1alpha1.ServiceImport:
-		return c.processors[ServiceImportsProcessorType]
-	case *discoveryv1.EndpointSlice:
-		return c.processors[EndpointSlicesProcessorType]
-	case *corev1.Secret:
-		return c.processors[SecretsProcessorType]
-	case *gwv1beta1.GatewayClass:
-		return c.processors[GatewayClassesProcessorType]
-	case *gwv1beta1.Gateway:
-		return c.processors[GatewaysProcessorType]
-	case *gwv1beta1.HTTPRoute:
-		return c.processors[HTTPRoutesProcessorType]
-	case *gwv1alpha2.GRPCRoute:
-		return c.processors[GRPCRoutesProcessorType]
-	case *gwv1alpha2.TCPRoute:
-		return c.processors[TCPRoutesProcessorType]
-	case *gwv1alpha2.TLSRoute:
-		return c.processors[TLSRoutesProcessorType]
-	case *gwpav1alpha1.RateLimitPolicy:
-		return c.processors[RateLimitPoliciesProcessorType]
-	case *gwpav1alpha1.SessionStickyPolicy:
-		return c.processors[SessionStickyPoliciesProcessorType]
-	case *gwpav1alpha1.LoadBalancerPolicy:
-		return c.processors[LoadBalancerPoliciesProcessorType]
-	case *gwpav1alpha1.CircuitBreakingPolicy:
-		return c.processors[CircuitBreakingPoliciesProcessorType]
-	}
-
-	return nil
-}
 
 func (c *GatewayCache) isRoutableService(service client.ObjectKey) bool {
 	for _, checkRoutableFunc := range []func(client.ObjectKey) bool{
